Revert "Merge remote-tracking branch 'origin/release/alpha/midterm-presentation' into feature/Item/Inventory"
This reverts commit 4973920657ca9d97dffaef61a177db82910fbf65, reversing
changes made to d93851bacc6debe0663df8481348a9f343c88b5f.

+++ LastCanary/Source/LastCanary/AI/BaseMonsterCharacter.h
@@ -153,7 +153,7 @@ protected:
     UAIPerceptionComponent* AIPerceptionComponent;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Monster|Combat")
-    float MaxHP;
+    int32 MaxHP;
 
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Monster|Combat")
     float AttackDamage;
